Refactor database_creation.py
